+++ .github/classroom/autograding.json
@@ -1,7 +1,7 @@
 {
   "tests": [
     {
-      "name": "Palindrome",
+      "name": "Tests",
       "setup": "npm install",
       "run": "npm test",
       "input": "",
